Started with hhds::tree<X> class structure and stub implementations
- The old tree.hpp file was moved to tree_pseudo.hpp.
- The current tree.hpp is the beginning of the project.
- Will test the current class after reviews on this commit
